change function
git-svn-id: https://vcs.maemo.org/svn/omweather/trunk@947 5aa3d57c-301e-0410-a7f2-8fec96c11689

+++ omweather/src/weather-window_popup.h
@@ -44,7 +44,7 @@ GtkWidget* create_24_hours_widget(int i, time_t current_time);
 GtkWidget* create_sun_time_widget(int i);
 void weather_window_popup_show(GtkWidget *widget,GdkEvent *event,gpointer user_data);
 void weather_window_settings(GtkWidget *widget,GdkEvent *event,gpointer user_data);
-int convert_wind_units(int to, int value);
+float convert_wind_units(int to, float value);
 /* extern functions */
 extern gpointer hash_table_find(gpointer key);
 extern void weather_initialize_dbus(void);
